adding relation user and post

+++ controller/feed.js
@@ -13,6 +13,7 @@ const path = require("path");
  * ========= Models ==============
  */
 const Post = require("../models/post");
+const User = require("../models/user");
 /**
  * ========= End Models ==============
  */
@@ -79,15 +80,15 @@ exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   const title = req.body.title;
   const content = req.body.content;
+  const userId = req.userId;
+  let creator;
   // For Windows User
   const imageUrl = req.file.path.replace("\\", "/");
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Anonim",
-    },
+    creator: userId,
   });
   if (!errors.isEmpty()) {
     const error = new Error("Validation Failed, entered data is incorrect.");
@@ -97,9 +98,21 @@ exports.createPost = (req, res, next) => {
   post
     .save()
     .then((result) => {
+      return User.findById(userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post Created!",
-        post: result,
+        post: post,
+        creator: {
+          _id: creator_id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -141,6 +154,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -174,8 +194,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({
